conflict

+++ inc/minirt.h
@@ -16,7 +16,7 @@
 # include <stdio.h>
 # include <unistd.h>
 # include <math.h>
-# include <float.h>
+# include <double.h>
 # include <stdbool.h>
 # include "../libft/libft.h" 
 # include "../minilibx-linux/mlx.h"
@@ -49,16 +49,16 @@ typedef struct	s_data
 
 typedef struct	s_vector
 {
-	float	a;
-	float	b;
-	float	c;
+	double	a;
+	double	b;
+	double	c;
 }	t_vector;
 
 typedef struct	s_pos
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_pos;
 
 typedef struct	s_color
@@ -80,12 +80,12 @@ typedef struct	s_light
 {
 	t_pos	pos;
 	t_color	color;
-	float	ratio;
+	double	ratio;
 }	t_light;
 
 typedef struct	s_ambient
 {
-	float	ratio;
+	double	ratio;
 	t_color	color;
 }	t_ambient;
 
@@ -101,15 +101,15 @@ typedef struct	s_cylinder
 	t_vector	vector;
 	t_pos		pos;
 	t_color		color;
-	float		d;
-	float		h;
+	double		d;
+	double		h;
 }	t_cylinder;
 
 typedef struct	s_sphere
 {
 	t_pos	pos;
 	t_color	color;
-	float	d;
+	double	d;
 }	t_sphere;
 
 // MAIN STRUCT -----------------------------
@@ -118,7 +118,7 @@ typedef struct	s_ray
 	t_vector	origin;
 	t_vector	direction;
 	t_color		color;
-	float		t;
+	double		t;
 	t_vector	hit_point;
 	t_vector	normal;
 	t_plane		*hit_plane;
@@ -150,8 +150,8 @@ typedef struct	s_pixel
 
 typedef	struct	s_world
 {
-	float		aspect_ratio;
-	float		fov_rad;
+	double		aspect_ratio;
+	double		fov_rad;
 	t_vector	right;
 	t_vector	up;
 	t_vector	forward;
@@ -179,11 +179,11 @@ void	intersection_sphere(t_params *params, t_ray *ray);
 void		free_all(t_params *params);
 t_vector	vector_add(t_vector v1, t_vector v2);
 t_vector	vector_sub(t_vector v1, t_vector v2);
-t_vector	vector_multi(float x, t_vector v1);
-t_vector	vector_divi(t_vector v1, float x);
+t_vector	vector_multi(double x, t_vector v1);
+t_vector	vector_divi(t_vector v1, double x);
 t_vector	vector_cross(t_vector v1, t_vector v2);
-float		vector_dot(t_vector v1, t_vector v2);
-float		vector_norm2(t_vector v1);
+double		vector_dot(t_vector v1, t_vector v2);
+double		vector_norm2(t_vector v1);
 void		vector_normalize(t_vector *v1);
 t_vector	pos_to_vector(t_pos pos);
